Разделила код на модули
